+++ common/src/main/java/earth/terrarium/adastra/client/dimension/MovementType.java
@@ -3,4 +3,5 @@ package earth.terrarium.adastra.client.dimension;
 public enum MovementType {
     STATIC,
     TIME_OF_DAY,
+    TIME_OF_DAY_REVERSED,
 }