+++ common/src/main/java/earth/terrarium/adastra/client/dimension/SkyRenderable.java
@@ -9,18 +9,19 @@ public record SkyRenderable(
     Vec3 globalRotation,
     Vec3 localRotation,
     MovementType movementType,
+    boolean blend,
     int backLightColor,
     float backLightScale) {
 
-    public static SkyRenderable create(ResourceLocation texture, float scale, Vec3 globalRotation, Vec3 localRotation, MovementType movementType) {
-        return new SkyRenderable(texture, scale, globalRotation, localRotation, movementType, 0, 0);
+    public static SkyRenderable create(ResourceLocation texture, float scale, Vec3 globalRotation, Vec3 localRotation, MovementType movementType, int backLightColor) {
+        return create(texture, scale, globalRotation, localRotation, movementType, false, backLightColor, scale * 3);
     }
 
-    public static SkyRenderable createWithBackLight(ResourceLocation texture, float scale, Vec3 globalRotation, Vec3 localRotation, MovementType movementType, int backLightColor) {
-        return createWithBackLight(texture, scale, globalRotation, localRotation, movementType, backLightColor, scale * 3);
+    public static SkyRenderable create(ResourceLocation texture, float scale, Vec3 globalRotation, Vec3 localRotation, MovementType movementType, boolean blend, int backLightColor) {
+        return create(texture, scale, globalRotation, localRotation, movementType, blend, backLightColor, scale * 3);
     }
 
-    public static SkyRenderable createWithBackLight(ResourceLocation texture, float scale, Vec3 globalRotation, Vec3 localRotation, MovementType movementType, int backLightColor, float backLightScale) {
-        return new SkyRenderable(texture, scale, globalRotation, localRotation, movementType, backLightColor, backLightScale);
+    public static SkyRenderable create(ResourceLocation texture, float scale, Vec3 globalRotation, Vec3 localRotation, MovementType movementType, boolean blend, int backLightColor, float backLightScale) {
+        return new SkyRenderable(texture, scale, globalRotation, localRotation, movementType, blend, backLightColor, backLightScale);
     }
 }