+++ common/src/main/java/earth/terrarium/adastra/client/dimension/renderers/ModSkyRenderer.java
@@ -4,7 +4,7 @@ import com.mojang.blaze3d.platform.GlStateManager;
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.*;
 import com.mojang.math.Axis;
-import earth.terrarium.adastra.client.dimension.MovementType;
+import earth.terrarium.adastra.client.dimension.ModDimensionSpecialEffects;
 import earth.terrarium.adastra.client.dimension.SkyRenderable;
 import earth.terrarium.adastra.client.utils.DimensionUtils;
 import earth.terrarium.adastra.mixins.client.LevelRendererAccessor;
@@ -13,8 +13,10 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.client.renderer.FogRenderer;
 import net.minecraft.client.renderer.GameRenderer;
+import net.minecraft.client.renderer.ShaderInstance;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.FastColor;
+import net.minecraft.util.Mth;
 import net.minecraft.util.RandomSource;
 import net.minecraft.util.random.SimpleWeightedRandomList;
 import net.minecraft.world.level.material.FogType;
@@ -27,6 +29,9 @@ import java.util.function.BiFunction;
 
 public class ModSkyRenderer {
     private final int stars;
+    private final int sunriseAngle;
+    private final boolean renderInRain;
+    private final int sunriseColor;
     private final BiFunction<ClientLevel, Float, Float> starBrightness;
     private final SimpleWeightedRandomList<Integer> starColors;
     private final List<SkyRenderable> skyRenderables;
@@ -36,11 +41,17 @@ public class ModSkyRenderer {
 
     public ModSkyRenderer(
         int stars,
+        int sunriseAngle,
+        boolean renderInRain,
+        int sunriseColor,
         BiFunction<ClientLevel, Float, Float> starBrightness,
         SimpleWeightedRandomList<Integer> starColors,
         List<SkyRenderable> skyRenderables
     ) {
         this.stars = stars;
+        this.sunriseAngle = sunriseAngle;
+        this.renderInRain = renderInRain;
+        this.sunriseColor = sunriseColor;
         this.starBrightness = starBrightness;
         this.starColors = starColors;
         this.skyRenderables = skyRenderables;
@@ -50,12 +61,16 @@ public class ModSkyRenderer {
         setupFog.run();
         if (isFoggy || inFog(camera)) return;
         if (starBuffer == null) createStars();
+        BufferBuilder bufferBuilder = Tesselator.getInstance().getBuilder();
 
         setSkyColor(level, camera, partialTick);
 
         RenderSystem.depthMask(false);
         VertexBuffer.unbind();
         RenderSystem.enableBlend();
+
+        renderSky(bufferBuilder, level, partialTick, poseStack, projectionMatrix);
+
         RenderSystem.blendFuncSeparate(GlStateManager.SourceFactor.SRC_ALPHA, GlStateManager.DestFactor.ONE, GlStateManager.SourceFactor.ONE, GlStateManager.DestFactor.ZERO);
         poseStack.pushPose();
 
@@ -63,19 +78,20 @@ public class ModSkyRenderer {
 
         RenderSystem.disableBlend();
 
-        BufferBuilder bufferBuilder = Tesselator.getInstance().getBuilder();
+        RenderSystem.setShaderColor(1, 1, 1, 1);
         skyRenderables.forEach(renderable -> {
-            var globalRotation = renderable.movementType() == MovementType.STATIC ?
-                renderable.globalRotation() :
-                renderable.globalRotation().add(0, 0, level.getTimeOfDay(partialTick) * 360 - 90);
+            var globalRotation = switch (renderable.movementType()) {
+                case STATIC -> renderable.globalRotation();
+                case TIME_OF_DAY -> renderable.globalRotation().add(0, 0, level.getTimeOfDay(partialTick) * 360);
+                case TIME_OF_DAY_REVERSED ->
+                    renderable.globalRotation().add(0, 0, -level.getTimeOfDay(partialTick) * 360);
+            };
 
-            renderSkyRenderable(bufferBuilder, poseStack, renderable.localRotation(), globalRotation, renderable.scale(), renderable.texture());
+            renderSkyRenderable(bufferBuilder, poseStack, renderable.localRotation(), globalRotation, renderable.scale(), renderable.texture(), renderable.blend());
             if (renderable.backLightScale() > 0) {
                 setSkyRenderableColor(level, partialTick, renderable.backLightColor());
-                RenderSystem.enableBlend();
-                renderSkyRenderable(bufferBuilder, poseStack, renderable.localRotation(), globalRotation, renderable.backLightScale(), DimensionUtils.BACKLIGHT);
+                renderSkyRenderable(bufferBuilder, poseStack, renderable.localRotation(), globalRotation, renderable.backLightScale(), DimensionUtils.BACKLIGHT, true);
                 RenderSystem.setShaderColor(1, 1, 1, 1);
-                RenderSystem.disableBlend();
             }
         });
 
@@ -102,7 +118,57 @@ public class ModSkyRenderer {
         RenderSystem.setShaderColor(r, g, b, 1);
     }
 
-    public void renderSkyRenderable(BufferBuilder bufferBuilder, PoseStack poseStack, Vec3 localRotation, Vec3 globalRotation, float scale, ResourceLocation texture) {
+    public void renderSky(BufferBuilder bufferBuilder, ClientLevel level, float partialTick, PoseStack poseStack, Matrix4f projectionMatrix) {
+        FogRenderer.levelFogColor();
+        ShaderInstance shader = RenderSystem.getShader();
+        if (shader == null) return;
+
+        var skyBuffer = ((LevelRendererAccessor) Minecraft.getInstance().levelRenderer).getSkyBuffer();
+        skyBuffer.bind();
+        skyBuffer.drawWithShader(poseStack.last().pose(), projectionMatrix, shader);
+        VertexBuffer.unbind();
+        RenderSystem.enableBlend();
+
+        float[] color = ModDimensionSpecialEffects.getSunriseColor(level.getTimeOfDay(partialTick), partialTick, sunriseColor);
+        if (color != null) {
+            renderSunrise(bufferBuilder, level, partialTick, poseStack, color);
+        }
+    }
+
+    public void renderSunrise(BufferBuilder bufferBuilder, ClientLevel level, float partialTick, PoseStack poseStack, float[] color) {
+        RenderSystem.setShader(GameRenderer::getPositionColorShader);
+        RenderSystem.setShaderColor(1, 1, 1, 1);
+
+        poseStack.pushPose();
+        poseStack.mulPose(Axis.XP.rotationDegrees(90));
+
+        poseStack.mulPose(Axis.ZP.rotationDegrees(sunriseAngle));
+
+        float sunAngle = Mth.sin(level.getSunAngle(partialTick)) < 0 ? 180 : 0;
+        poseStack.mulPose(Axis.ZP.rotationDegrees(sunAngle));
+        poseStack.mulPose(Axis.ZP.rotationDegrees(90));
+
+        float r = color[0];
+        float g = color[1];
+        float b = color[2];
+
+        Matrix4f matrix = poseStack.last().pose();
+        bufferBuilder.begin(VertexFormat.Mode.TRIANGLE_FAN, DefaultVertexFormat.POSITION_COLOR);
+        bufferBuilder.vertex(matrix, 0, 100, 0).color(r, g, b, color[3]).endVertex();
+
+        for (int i = 0; i <= 16; i++) {
+            float angle = (float) i * (float) (Math.PI * 2) / 16;
+            float x = Mth.sin(angle);
+            float y = Mth.cos(angle);
+            bufferBuilder.vertex(matrix, x * 120, y * 120, -y * 40 * color[3]).color(r, g, b, 0).endVertex();
+        }
+
+        BufferUploader.drawWithShader(bufferBuilder.end());
+        poseStack.popPose();
+    }
+
+    public void renderSkyRenderable(BufferBuilder bufferBuilder, PoseStack poseStack, Vec3 localRotation, Vec3 globalRotation, float scale, ResourceLocation texture, boolean blend) {
+        if (blend) RenderSystem.enableBlend();
         poseStack.pushPose();
 
         poseStack.mulPose(Axis.XP.rotationDegrees((float) globalRotation.x));
@@ -132,8 +198,8 @@ public class ModSkyRenderer {
         float r = FastColor.ARGB32.red(color) / 255f;
         float g = FastColor.ARGB32.green(color) / 255f;
         float b = FastColor.ARGB32.blue(color) / 255f;
-//        float a = 1 - level.getRainLevel(partialTick); // TODO
-        RenderSystem.setShaderColor(r, g, b, 1);
+        float a = renderInRain ? 1 : 1 - level.getRainLevel(partialTick);
+        RenderSystem.setShaderColor(r, g, b, a);
     }
 
     public void renderStars(ClientLevel level, float partialTick, PoseStack poseStack, Matrix4f projectionMatrix, Runnable setupFog) {