+++ common/src/main/java/earth/terrarium/adastra/client/utils/DimensionUtils.java
@@ -10,6 +10,7 @@ public class DimensionUtils {
 
     public static final ResourceLocation SUN = new ResourceLocation(AdAstra.MOD_ID, "textures/environment/sun.png");
     public static final ResourceLocation BLUE_SUN = new ResourceLocation(AdAstra.MOD_ID, "textures/environment/blue_sun.png");
+    public static final ResourceLocation RED_SUN = new ResourceLocation(AdAstra.MOD_ID, "textures/environment/red_sun.png");
 
     public static final ResourceLocation EARTH = new ResourceLocation(AdAstra.MOD_ID, "textures/environment/earth.png");
     public static final ResourceLocation MOON = new ResourceLocation(AdAstra.MOD_ID, "textures/environment/moon.png");
@@ -17,6 +18,10 @@ public class DimensionUtils {
     public static final ResourceLocation VENUS = new ResourceLocation(AdAstra.MOD_ID, "textures/environment/venus.png");
     public static final ResourceLocation MERCURY = new ResourceLocation(AdAstra.MOD_ID, "textures/environment/mercury.png");
     public static final ResourceLocation GLACIO = new ResourceLocation(AdAstra.MOD_ID, "textures/environment/glacio.png");
+
+    public static final ResourceLocation PHOBOS = new ResourceLocation(AdAstra.MOD_ID, "textures/environment/phobos.png");
+    public static final ResourceLocation DEIMOS = new ResourceLocation(AdAstra.MOD_ID, "textures/environment/deimos.png");
+    public static final ResourceLocation VICINUS = new ResourceLocation(AdAstra.MOD_ID, "textures/environment/vicinus.png");
 
     public static int getTicks() {
         return ((LevelRendererAccessor) Minecraft.getInstance().levelRenderer).getTicks();