+++ common/src/main/java/earth/terrarium/adastra/mixins/client/LevelRendererAccessor.java
@@ -1,5 +1,6 @@
 package earth.terrarium.adastra.mixins.client;
 
+import com.mojang.blaze3d.vertex.VertexBuffer;
 import net.minecraft.client.Camera;
 import net.minecraft.client.renderer.LevelRenderer;
 import org.spongepowered.asm.mixin.Mixin;
@@ -14,4 +15,7 @@ public interface LevelRendererAccessor {
 
     @Accessor
     int getTicks();
+
+    @Accessor
+    VertexBuffer getSkyBuffer();
 }