+++ changelog.md
@@ -84,6 +84,7 @@ should now be clear when you want to land on the planet, land on the space stati
 You must build a space station and land on that space station first.
 - Added industrial lamps and small industrial lamps. These are available in all 16 colors.
 - Updated star colors to be more vibrant and have some colors be more rare than others.
+- Improved planet skyboxes
 
 ## Breaking Changes
 - storage blocks like desh blocks, steel blocks, etc. have had an ID change. Turn them into ingots before